Force-framework: renamed view package.
