+++ src/main/java/br/com/schumaker/force/framework/ioc/reflection/validation/EmailValidation.java
@@ -1,9 +1,9 @@
-package br.com.schumaker.octopus.framework.ioc.reflection.validation;
+package br.com.schumaker.force.framework.ioc.reflection.validation;
 
-import br.com.schumaker.octopus.framework.annotations.validations.Email;
-import br.com.schumaker.octopus.framework.annotations.validations.Validate;
-import br.com.schumaker.octopus.framework.exception.OctopusException;
-import br.com.schumaker.octopus.framework.web.http.Http;
+import br.com.schumaker.force.framework.annotations.validations.Email;
+import br.com.schumaker.force.framework.annotations.validations.Validate;
+import br.com.schumaker.force.framework.exception.ForceException;
+import br.com.schumaker.force.framework.web.http.Http;
 
 import java.lang.reflect.Field;
 
@@ -26,10 +26,10 @@ public class EmailValidation implements Validation {
      *
      * @param object the object to be validated.
      * @param field the field to be validated.
-     * @throws OctopusException if the field is not a valid email address.
+     * @throws ForceException if the field is not a valid email address.
      */
     @Override
-    public void validate(Object object, Field field) throws OctopusException {
+    public void validate(Object object, Field field) throws ForceException {
         try {
             field.setAccessible(true);
             Object fieldValue = field.get(object);
@@ -37,13 +37,13 @@ public class EmailValidation implements Validation {
                 if (!EmailValidator.isValidEmail(email)) {
                     var message = field.getAnnotation(Email.class).value();
                     if (message.equals(EMAIL_VALIDATION_MESSAGE)) {
-                        throw new OctopusException(String.format(message, field.getName()), Http.HTTP_400);
+                        throw new ForceException(String.format(message, field.getName()), Http.HTTP_400);
                     }
-                    throw new OctopusException(message);
+                    throw new ForceException(message);
                 }
             }
         } catch (IllegalAccessException e) {
-            throw new OctopusException("Failed to validate field " + field.getName(), e);
+            throw new ForceException("Failed to validate field " + field.getName(), e);
         }
     }
 }
