+++ src/main/java/br/com/schumaker/force/framework/ioc/reflection/validation/EmailValidator.java
@@ -1,4 +1,4 @@
-package br.com.schumaker.octopus.framework.ioc.reflection.validation;
+package br.com.schumaker.force.framework.ioc.reflection.validation;
 
 import java.util.regex.Pattern;
