+++ src/main/java/br/com/schumaker/force/framework/web/view/PageRequest.java
@@ -1,6 +1,6 @@
-package br.com.schumaker.octopus.framework.web.view;
+package br.com.schumaker.force.framework.web.view;
 
-import br.com.schumaker.octopus.framework.jdbc.Sort;
+import br.com.schumaker.force.framework.jdbc.Sort;
 
 /**
  * The PageRequest class.
